+++ src/lib/utils.js
@@ -28,7 +28,7 @@ export const logger = {
  * Converts the arguments array into a readable object.
  * @returns {Object} An object containing the command line arguments.
  */
-export const getArguments = () => process.argv.reduce((acc, arg) => {
+export const getArguments = () => process.argv.slice(3).reduce((acc, arg) => {
   if (arg.includes('=') || /^-+\w+/.test(arg)) {
     const { 0: key, 1: value } = arg.split('=')
     return value === 'undefined' ? acc : {
